backend/domains/student_management/user-get-profile-service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
psycopg2-binary
